modifications to work with new sim
