+++ compspec.json
@@ -1,24 +1,25 @@
 {
   "meta": {
-    "name": "single shot regression demo",
+    "name": "coinstac first example",
     "id": "avg_test",
     "version": "v1.0.0",
     "repository": "https:\/\/github.com\/trendscenter\/coinstac-first-example",
-    "description": "a test from single shot regresssion"
+    "description": "An example code to demonstrate coinstac's working",
+    "compspecVersion": 2
   },
   "computation": {
     "type": "docker",
-    "dockerImage": "avg_test",
+    "dockerImage": "coinstacteam/coinstac-example",
     "command": [
       "python",
-      "\/computation\/local.py"
+      "\/computation\/scripts\/local.py"
     ],
     "remote": {
       "type": "docker",
-      "dockerImage": "avg_test",
+      "dockerImage": "coinstacteam/coinstac-example",
       "command": [
         "python",
-        "\/computation\/remote.py"
+        "\/computation\/scripts\/remote.py"
       ]
     },
     "input": {
