+++ requirements.txt
@@ -1,1 +1,2 @@
+coinstac==0.3.3b0
 numpy
