+++ scripts/entry.py
@@ -1,0 +1,5 @@
+import coinstac
+import scripts.local as local
+import scripts.remote as remote
+
+coinstac.start(local.start, remote.start)
